requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.1.1
openpyxl==3.1.2
tabula-py==2.8.2